fixup! refactor!: update scalar challenges to use mask
