+++ crates/proof-of-sql/src/base/scalar/scalar.rs
@@ -2,6 +2,7 @@
 
 use crate::base::{encode::VarInt, ref_into::RefInto, scalar::ScalarConversionError};
 use alloc::string::String;
+use bnum::types::U256;
 use core::ops::Sub;
 use num_bigint::BigInt;
 
@@ -73,28 +74,5 @@ pub trait Scalar:
     const TEN: Self;
     /// The value to mask the challenge with to ensure it is in the field.
     /// This one less than the largest power of 2 that is less than the field modulus.
-    const CHALLENGE_MASK: [u64; 4];
-}
-
-#[cfg(test)]
-pub(crate) fn test_scalar_constants<S: Scalar>() {
-    assert_eq!(S::from(0), S::ZERO);
-    assert_eq!(S::from(1), S::ONE);
-    assert_eq!(S::from(2), S::TWO);
-    // -1/2 == least upper bound
-    assert_eq!(-S::TWO.inv().unwrap(), S::MAX_SIGNED);
-    assert_eq!(S::from(10), S::TEN);
-
-    // Check the challenge mask
-    let mid_point_limbs: [u64; 4] = S::MAX_SIGNED.into();
-    let modulus_minus_one_limbs: [u64; 4] = (-S::ONE).into();
-    assert_eq!(S::CHALLENGE_MASK[0], u64::MAX);
-    assert_eq!(S::CHALLENGE_MASK[1], u64::MAX);
-    assert_eq!(S::CHALLENGE_MASK[2], u64::MAX);
-    assert_eq!(
-        S::CHALLENGE_MASK[3],
-        u64::MAX >> S::CHALLENGE_MASK[3].leading_zeros()
-    );
-    assert!(mid_point_limbs[3] < S::CHALLENGE_MASK[3]);
-    assert!(modulus_minus_one_limbs[3] > S::CHALLENGE_MASK[3]);
+    const CHALLENGE_MASK: U256;
 }
